+++ example/lib/main.dart
@@ -3,6 +3,7 @@
 import 'package:flutter/material.dart';
 
 import 'main_screen.dart';
+import 'demos/bluetooth_scanner.dart' show BluetoothScannerDemo;
 import 'demos/face_detector.dart' show FaceDetectorDemo;
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -39,6 +40,7 @@ class _AppState extends State<App> {
       ),
       home: MainScreen(),
       routes: <String, WidgetBuilder>{
+        '/bluetooth_scanner': (context) => BluetoothScannerDemo(),
         '/face_detection': (context) => FaceDetectorDemo(),
       },
     );